updating skeleton to use pagination and moving files to correct location
